Moved traj functions into own file
